docker/learntrac/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary,pool]==3.1.12
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7